requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
python-dotenv==0.19.0
